全体的に処理を変更
