+++ config/common.json
@@ -1,0 +1,6 @@
+{
+    "port": {
+        "kotori": 27133,
+        "mahiru": 27132
+    }
+}
